Display inactive accounts, show updated card descriptions in filters
