+++ src/react/Components/FilterComponents/CardSelection.jsx
@@ -17,6 +17,7 @@ import FilterListIcon from "@material-ui/icons/FilterList";
 import { cardsUpdate } from "../../Actions/cards";
 import { addCardIdFilter, removeCardIdFilter, toggleCardIdFilter } from "../../Actions/filters";
 import { getCardTypeImage } from "../../Pages/Cards/CardListItem";
+import { getCardDescription } from "../../Functions/Utils";
 
 const styles = {
     listItem: {
@@ -100,12 +101,15 @@ class CardSelection extends React.Component {
         });
 
         const cardMenuItems = Object.keys(cards)
-            .filter(cardIndex => {
-                const card = cards[cardIndex];
-                if (card && card.status !== "ACTIVE") {
-                    return false;
-                }
-                return true;
+            // put inactive cards at the bottom
+            .sort((index1, index2) => {
+                const card1 = cards[index1];
+                const card2 = cards[index2];
+
+                if (card1.status !== "ACTIVE" && card2.status === "ACTIVE") return 1;
+                if (card1.status === "ACTIVE" && card2.status !== "ACTIVE") return -1;
+
+                return 0;
             })
             .map((cardIndex, key) => {
                 const card = cards[cardIndex];
@@ -114,6 +118,8 @@ class CardSelection extends React.Component {
                 if (selectedCardIds.includes(card.id)) {
                     return null;
                 }
+
+                const cardDescription = getCardDescription(card);
                 const { cardImage, cardType } = getCardTypeImage(card.type, card.cardType);
 
                 return (
@@ -121,7 +127,7 @@ class CardSelection extends React.Component {
                         <ListItemIcon style={styles.listItemIcon}>
                             <img style={styles.listItemImg} src={cardImage} />
                         </ListItemIcon>
-                        {card.second_line || cardType}
+                        {cardDescription || cardType}
                     </MenuItem>
                 );
             });
